refactor: rename filenames using kebab-case
