+++ project-3/hooks/use-fetch-weather.ts
