+++ project-3/jest.setup.js
@@ -1,5 +1,5 @@
 import '@testing-library/jest-dom/extend-expect';
-import { server } from './__api_mocks__/mswServer';
+import { server } from './__api_mocks__/msw-server';
 
 beforeAll(() => server.listen());
 afterEach(() => server.resetHandlers());
